added class & init
